Updating version number and data formatter.

+++ DDComet/ISO8601DateFormatter.m
@@ -686,7 +686,7 @@ static BOOL is_leap_year(NSUInteger year);
 		if (offset == 0)
 			str = [str stringByAppendingString:ISO_TIMEZONE_UTC_FORMAT];
 		else
-			str = [str stringByAppendingFormat:ISO_TIMEZONE_OFFSET_FORMAT, offset / 60, offset % 60];
+			str = [str stringByAppendingFormat:ISO_TIMEZONE_OFFSET_FORMAT, (long)(offset / 60), (long)(offset % 60)];
 	}
 
 	//Undo the change we made earlier
